Moving Brave custom components and styles in to ui/
