+++ ui/app/components/app/connect-wallet/index.scss
