+++ ui/app/components/app/password-warning/index.scss
