+++ ui/app/pages/welcome/index.scss
