+++ ui/app/pages/welcome/welcome.container.js
